requirements.txt
boto3 == 1.24.4
requests2 == 0.2
logging == 1.1.1
logger == 0.3.3